refactor file paths
