+++ apps/frontend/src/components/MockUsers.tsx
@@ -1,4 +1,4 @@
-export let mockUsers = [
+export const mockUsers = [
   { id: 1, username:"mock1", email: 'mock1@example.com', password: 'password1' },
   { id: 2, username:"mock2", email: 'mock2@example.com', password: 'password2' },
 ]
