+++ apps/frontend/src/components/Navbar.tsx
