+++ apps/frontend/src/components/ValidatedInput.tsx
