+++ apps/frontend/src/components/enums/Category.tsx
