+++ apps/frontend/src/hook/useQuestion.ts
@@ -1,7 +1,7 @@
 import React, { useEffect } from 'react'
 import { Question } from '../components/Question';
 import { mockQuestions } from '../components/MockQuestions';
-import { getAllQuestions } from "../src/utils/database/question/Question";
+import { getAllQuestions } from "../utils/database/question/Question";
 
 function useQuestion() {
     const [questions, setQuestions] = React.useState<Question[]>(mockQuestions);
