+++ apps/frontend/src/pages/_app.tsx
@@ -2,7 +2,7 @@ import React, { useEffect } from "react";
 import Head from "next/head";
 import { AppProps } from "next/app";
 import "bootstrap/dist/css/bootstrap.min.css";
-import "../styles/globals.css";
+import "/styles/globals.css";
 import Layout from "../components/Layout";
 import { SessionProvider } from "next-auth/react";
