+++ apps/frontend/src/pages/api/auth/[...nextauth].ts
